Second Commit

+++ src/utils/preprocessing_utils.py
@@ -3,5 +3,5 @@ import pandas as pd
 def data_split(dir):
     df = pd.read_csv(dir)
     X = df.drop(columns='phishing')
-    y = df['phishing']
+    y = df['phishing'].astype('int')
     return X,y
